rename de.torui.coflsky.core to commands in for refactor
