+++ src/main/java/de/torui/coflsky/commands/JsonStringCommand.java
@@ -1,4 +1,4 @@
-package de.torui.coflsky.core;
+package de.torui.coflsky.commands;
 
 import com.google.gson.GsonBuilder;
 import com.google.gson.reflect.TypeToken;
